requirements.txt
art==5.5
asciiart==0.2
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
Pillow==9.0.1
requests==2.27.1
soupsieve==2.3.1
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
python-dotenv==0.19.2